requirements.txt
google-adk==1.2.1
a2a-sdk==0.2.5
fastapi==0.115.12
litellm==1.72.0
uvicorn[standard]==0.34.3
streamlit==1.45.1
python-dotenv==1.1.0
httpx==0.28.1
toolbox-core==0.3.0